+++ index.js
@@ -1,17 +1,20 @@
 module.exports = function(pattern, params) {
   var result = pattern;
 
+  // First, replace non-glob params (:xxx params)
   for(var param in params) {
-    var array_index = params[param] == params[+param];
-    if(!array_index) {
-      result = result.replace(new RegExp(':' + param), encodeURIComponent(params[param]));
-    }
+    result = result.replace(':' + param, encodeURIComponent(params[param]));
   }
 
+  // Second, replace glob params (*)
   if(Array.isArray(params)) {
     for(var n = 0; n < params.length; n++) {
       result = result.replace('*', encodeURIComponent(params[n]));
     }
+  }
+
+  if(result.match(/[\*:]/)) {
+    throw new Error('There were unexpanded params: ' + result);
   }
 
   return result;