+++ History.md
@@ -1,3 +1,8 @@
+## 0.1.0
+
+* Fail when there are unexpanded params.
+* Test compatibility with express routes.
+
 ## 0.0.2
 
 * Publish as bower package.